+++ SortedMenus/SortedMenus/CraftingMenu/CraftingMenuSorting.cs
@@ -1,38 +1,18 @@
-﻿using HarmonyLib;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
 namespace SortedMenus
 {
-    [HarmonyPatch]
-    internal class CraftingMenuPatches
+    internal class CraftingMenuSorting
     {
-        internal const string handCrafting = "hand crafting";
-
-        [HarmonyPriority(Priority.VeryLow)]
-        [HarmonyPatch(typeof(InventoryGui), nameof(InventoryGui.UpdateRecipeList)), HarmonyPrefix]
-        private static void UpdateRecipeList(ref List<Recipe> recipes)
+        internal static void UpdateRecipeList(ref List<Recipe> recipes, string stationName, bool isHandCrafting)
         {
-            if (!Player.m_localPlayer)
-            {
-                return;
-            }
-
             if (SortConfig.CraftingMenuSorting.Value == SortCraftingMenu.Disabled)
             {
                 return;
             }
 
-            CraftingStation currentCraftingStation = Player.m_localPlayer.GetCurrentCraftingStation();
-
             // no need to sort the hand crafting menu while not in debug mode
-            if (!currentCraftingStation && !Player.m_localPlayer.NoCostCheat())
-            {
-                return;
-            }
-
-            string stationName = !currentCraftingStation ? handCrafting : currentCraftingStation.m_name;
-
-            if (stationName == "$piece_cauldron")
+            if (isHandCrafting && !Player.m_localPlayer.NoCostCheat())
             {
                 return;
             }
@@ -40,8 +20,13 @@ namespace SortedMenus
             List<Recipe> relevantCache = null;
             CraftingStationType stationType = CraftingStationType.Other;
 
-            if (stationName == "$piece_forge")
+            if (isHandCrafting)
             {
+                relevantCache = SortedRecipeCaches.cachedSortedNoCostRecipes;
+            }
+            else if (stationName == "forge")
+            {
+                // the prefab key is not "piece_forge" for some reason, even if the translation key is
                 stationType = CraftingStationType.Forge;
 
                 if (Player.m_localPlayer.NoCostCheat())
@@ -53,7 +38,7 @@ namespace SortedMenus
                     relevantCache = SortedRecipeCaches.cachedSortedForgeRecipes;
                 }
             }
-            else if (stationName == "$piece_workbench")
+            else if (stationName == "piece_workbench")
             {
                 if (Player.m_localPlayer.NoCostCheat())
                 {
@@ -63,10 +48,6 @@ namespace SortedMenus
                 {
                     relevantCache = SortedRecipeCaches.cachedSortedWorkBenchRecipes;
                 }
-            }
-            else if (stationName == handCrafting && Player.m_localPlayer.NoCostCheat())
-            {
-                relevantCache = SortedRecipeCaches.cachedSortedNoCostRecipes;
             }
 
             var sw = new System.Diagnostics.Stopwatch();